fix(review): address code review findings
- ForwardedHeaders: ForwardLimit=1 + KnownIPNetworks/KnownProxies.Clear() so
  Fly proxy XFF is applied and RateLimitFilter gets per-client IPs
- RateLimitFilter: add periodic sweep to evict expired buckets (memory leak)
- ListLinks: page/pageSize are now optional (int?) with defaults; page clamped
  to [1,10_000] and OFFSET cast to long to prevent int overflow -> 500
- Db.NormalizeUri: apply Uri.UnescapeDataString to username/password/database
  so %-encoded Neon postgres:// passwords are handled correctly
- FollowLink: click recording is now best-effort (try/catch) so an analytics
  failure cannot block the redirect response
- Program: SeedApiKey and BaseUrl are now required (throw if unset) to prevent
  silent seeding with a known weak key in production
- ApiKeyValidator: return hash (= OwnerId) from ValidateAsync to eliminate the
  redundant second SHA-256 call in ApiKeyAuthFilter
- Tests: add SecondApiKey fixture; add tenant-isolation tests for stats/delete/list
  (previously the wrong-owner 404 branch was never exercised); add ListLinks
  no-query-string -> 200 test; regenerate openapi.json and SliceApiClient.g.cs

+++ src/ShortLink.Api/Infrastructure/PostgresLinkStore.cs
@@ -68,7 +68,8 @@ public sealed class PostgresLinkStore(NpgsqlDataSource db) : ILinkStore
             """;
         cmd.Parameters.AddWithValue(ownerKeyHash);
         cmd.Parameters.AddWithValue(pageSize);
-        cmd.Parameters.AddWithValue((page - 1) * pageSize);
+        // Cast to long before multiplication to prevent int overflow for large page values.
+        cmd.Parameters.AddWithValue((long)(page - 1) * pageSize);
         await using var reader = await cmd.ExecuteReaderAsync(ct);
         var results = new List<LinkRecord>();
         while (await reader.ReadAsync(ct))
